frontend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.8.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.11.1
anyio==4.6.2.post1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.3
attrs==24.2.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.15.1
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dnspython==2.2.1
docstring_parser==0.16
email-validator==1.3.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==0.8.3
fastapi==0.78.0
fastjsonschema==2.16.1
filelock==3.16.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.37.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
googletrans==3.0.0
greenlet==1.1.3
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==1.0.6
httptools==0.4.0
httpx==0.27.2
hyperframe==5.2.0
identify==2.6.1
idna==3.10
importlib-metadata==4.12.0
iniconfig==2.0.0
instructor==1.5.2
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.2.0
json5==0.9.8
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
kiwisolver==1.4.4
Mako==1.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
multidict==6.1.0
narwhals==1.17.0
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.6
nodeenv==1.9.1
notebook-shim==0.1.0
npm==0.1.1
numpy==1.23.5
openai==1.51.2
optional-django==0.1.0
packaging==24.1
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
pre_commit==4.0.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
propcache==0.2.0
protobuf==5.29.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==18.1.0
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.0.2
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==8.3.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.1
PyYAML==6.0.2
pyzmq==23.2.0
querycsv==0.3.0.0
requests==2.32.3
rfc3986==1.4.0
rich==13.9.2
scikit-learn==1.1.2
scipy==1.9.1
Send2Trash==1.8.0
shellingham==1.5.4
six==1.16.0
sklearn==0.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
stack-data==0.3.0
starlette==0.19.1
streamlit==1.40.2
swarm @ git+https://github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47
tenacity==8.5.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tornado-sqlalchemy==0.7.0
tqdm==4.66.5
traitlets==5.3.0
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.26.6
watchgod==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.2.2
WTForms==3.0.1
xgboost==2.1.2
yarl==1.15.2
zipp==3.8.1
